update build script and other minor changes
remove some hard code string
